Improve maskAbsentMethods interaction with new sequence stuff
