+++ src/errors/NotBoundedError.js
@@ -1,6 +1,4 @@
-import {sequenceConverters} from "../core/asSequence";
 import {error} from "../core/error";
-import {isSequence} from "../core/sequence";
 import {isArray, isString} from "../core/types";
 
 export const NotBoundedError = error({
@@ -19,14 +17,11 @@ export const NotBoundedError = error({
     constructor: function NotBoundedError(source, options = undefined){
         this.source = source;
         this.options = options || {};
-        const knownBounded = (isSequence(source) && source.unbounded() ?
+        const knownBounded = (source.unbounded() ?
             "known to be unbounded" : "not known to be bounded"
         );
-        const sourceType = (isSequence(source) ?
-            "a " + source.typeChainString() + " sequence" : "an iterable"
-        );
         this.message = (
-            `The action requires fully consuming ${sourceType} that is ` +
+            `The action requires fully consuming a sequence that is ` +
             `${knownBounded}. ` +
             "Try using a method such as 'head', 'limit', or 'assumeBounded' to " +
             "resolve this error."
@@ -40,21 +35,7 @@ export const NotBoundedError = error({
         }
     },
     enforce: function(source, options){
-        if(
-            (isSequence(source) && source.bounded()) ||
-            isArray(source) || isString(source)
-        ){
-            return source;
-        }
-        for(const converter of sequenceConverters){
-            if(converter.predicate(source)){
-                if(converter.bounded(source)){
-                    return source;
-                }else{
-                    break;
-                }
-            }
-        }
+        if(source.bounded()) return source;
         throw NotBoundedError(source, options);
     },
 });
